+++ packages/accordion/src/Accordion/styles.ts
@@ -1,5 +1,5 @@
-import { styled } from '@siakit/core'
 import * as RadixAccordion from '@radix-ui/react-accordion'
+import { styled } from '@siakit/core'
 
 export const AccordionContainer = styled(RadixAccordion.Root, {
   flex: 1,