+++ packages/accordion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @siakit/accordion
 
+## 0.0.5
+
+### Patch Changes
+
+- fix accordion props
+
 ## 0.0.4
 
 ### Patch Changes